// src/ui/interface.ts
export interface PageSettings {
    id?: string;
    is_active: boolean;
    title: string;
    slug: string;
    content: string;
    created_by: string;
    created_at: Date;
    updated_by: string;
    updated_at: Date;
}

export interface PageUpdate {
    id?: string;
    is_active: boolean;
    title: string;
    slug: string;
    content: string;
    updated_by: string;
    updated_at: Date;
}

export interface PostSettings extends PageSettings { // use in create flow
    cover_image_url: string;
    category: string[];
}

export interface PostsData extends PageSettings { // use in fetch flow
    cover_image_url: string;
    category: string;
}

// SETTINGS 

export interface SettingsData {
    name: string;
    data: string;
} 

export interface Settings {
    mainpage: SettingsData;
    notification: SettingsData;
    announcement: SettingsData;
    herosection: SettingsData;
    [key: string]: SettingsData;
}
export interface MainPageSettings {
    [key: string]: MainPageData[];
}

export interface MainPageData {
    title: string;
    route: string;
}

export interface NotificationSettings {
    title: string;
    message: string;
    route: string;
    createdBy: string;
    createdAt: Date;
    expiredAt: Date;
}

export interface AnnouncementSettings {
    title: string;
    img_url: string;
    route: string;
    createdBy: string;
    createdAt: Date;
    expiredAt: Date;
}

export interface HeroSectionData {
    title: string;
    img_url: string;
    route: string;
}

export interface HeroSectionSettings {
    posts_show_length : number;
    data: HeroSectionData[];
}